fold the minimal program into harnesses.utils
core.py now carries the argument parsing, assembly, and __main__
entry point - inert on package import, the entry point once the
module ends the bundled script. The program text is reduced to the
PEP 723 metadata block, built as CHAT_PROGRAM_SOURCE in launch.py,
and the harnesses.minimal package is gone.
