+++ verifiers/v1/harnesses/null/harness.py
@@ -1,8 +1,10 @@
 from verifiers.v1.clients import ModelContext
 from verifiers.v1.configs.harness import HarnessConfig
 from verifiers.v1.harness import Harness
-from verifiers.v1.harnesses.minimal import PROGRAM_SOURCE
-from verifiers.v1.harnesses.utils.launch import launch_chat_program
+from verifiers.v1.harnesses.utils.launch import (
+    CHAT_PROGRAM_SOURCE,
+    launch_chat_program,
+)
 from verifiers.v1.runtimes import ProgramResult, Runtime
 from verifiers.v1.task import TaskData
 from verifiers.v1.trace import Trace
@@ -20,7 +22,7 @@ class NullHarness(Harness[NullHarnessConfig]):
     NEEDS_CONTAINER = False
 
     async def setup(self, runtime: Runtime) -> None:
-        await runtime.prepare_uv_script(PROGRAM_SOURCE, self.config.resolved_env)
+        await runtime.prepare_uv_script(CHAT_PROGRAM_SOURCE, self.config.resolved_env)
 
     async def launch(
         self,
@@ -34,7 +36,7 @@ class NullHarness(Harness[NullHarnessConfig]):
     ) -> ProgramResult:
         system_prompt, prompt = self.resolve_prompt(data)
         return await launch_chat_program(
-            PROGRAM_SOURCE,
+            CHAT_PROGRAM_SOURCE,
             self.config,
             ctx,
             trace,
